requirements.txt
langgraph
langchain-google-genai
langchain-community
langchain-huggingface
firecrawl
google-generativeai
pydantic
requests
python-dotenv
faiss-cpu
unstructured
markdown
sentence-transformers
transformers
accelerate
torch
nltk
scikit-fuzzy
tiktoken 
pypdf
pandas
matplotlib
seaborn